+++ TestProject/core/src/Objects/BulletFormation/TargetDownwardLinearBulletFormation.java
@@ -17,7 +17,7 @@ public class TargetDownwardLinearBulletFormation extends BulletFormation {
     }
 
     @Override
-    public List<Ammo> shoot(Entity entity, int xShootPosition, int yShootPosition) {
+    public List<Ammo> shoot(Entity entity, float xShootPosition, float yShootPosition) {
         // Set up multiply
         float playerXPosition = player.getXPosition();
         float playerYPosition = player.getYPosition();
